requirements.txt
# Core async/HTTP
aiohttp>=3.9.0
aiofiles>=23.2.0

# Environment/config
python-dotenv>=1.0.0

# Data analysis
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# OpenAI-compatible client (for Lambda.ai)
openai>=1.0.0
